legacy filetag support
filetag has changed drastically since the recent refactor that went live in
the first week of Nov 2022. Rethink apps on version v053i,j,k though do not
understand this new filetag version. Previously, the 16 Nov 2022 update for
codec u8 blocklists was planned to be the final for app versions v053k, but
now the plan is to support for a forseeable future.

The blocklists built after 23 Nov 2022 come with a backwards-compatible
filetag-legacy.json. For codec u8, this legacy filetag is the default.
For codec u6, the newer filetag.json is the default.